adjust configs

+++ backend/src/api.py
@@ -34,7 +34,7 @@ def home():
   return response
 
 @app.post("/text_grade/")
-async def text_grade(request: Request) -> dict[str, int]:
+async def text_grade(request: Request) -> dict[str, float]:
   response = {
     "grade": predict_from_text(request.essay)
   }
